File scan node explain string value correction. Use consistent hash to collect BE only when the file cache is enabled.
